feat: wire Model Plane agentic browser loop to Quarry (real chromiumoxide + LLM)
Connects execution-core's browser-agent loop to Quarry's real chromiumoxide
browser via a new /v1/agent/* endpoint, with next-action selection driven by a
real inference-core LLM call. No mocks.

Quarry-edge (crates/quarry-edge):
- agent_routes.rs: POST /v1/agent/runs, POST /v1/agent/runs/{id}/step,
  DELETE /v1/agent/runs/{id}. Leased browser session per run (run->session
  map), per-step ObservationRunner, AgentStarted/AgentCompleted events,
  caller-supplied profile reuse.
- state.rs/main.rs/routes.rs/lib.rs/Cargo.toml: AppState gains agent_driver
  (real ChromiumoxideDriver behind default `browser-agent` feature) + agent_runs;
  routes registered under require_auth.
- clippy -D warnings cleanups (audio_routes, firecrawl_adapter, routes).

Model Plane (rust/services/execution-core):
- quarry_agent.rs: HTTP client for /v1/agent/*, envelope unwrap, internal<->wire
  action/observation mapping.
- llm_planner.rs: real inference-core gRPC (InferenceCore.Infer) with a
  JSON-schema-constrained next action; gated by QUARRY_BROWSER_AGENT_LLM.
- browser_agent.rs: async loop start_run -> step -> observe -> LLM decide ->
  close_run; domain allow-list enforced at dispatch; ZDR + max_cost_usd plumbed.
- runtime_loop/tool_bridge: async browser_agent dispatch path.
- tests/browser_agent_e2e.rs: real end-to-end (ignored; needs live stack).

Env: QUARRY_BROWSER_AGENT_ENABLED, QUARRY_EDGE_URL, QUARRY_EDGE_TOKEN,
QUARRY_BROWSER_AGENT_LLM, INFERENCE_CORE_URL, QUARRY_BROWSER_AGENT_MODEL.

Plan: docs/plans/quarry-browser-agent-loop-plan.md

+++ apps/Ingestion Plane/Quarry-v2/crates/quarry-edge/src/firecrawl_adapter.rs
@@ -92,7 +92,7 @@ pub fn adapt(output: &NormalizedOutput) -> FirecrawlResponse {
                 description: None,
                 language: output.metadata.lang.clone(),
                 source_url: output.url.requested.clone(),
-                status_code: output.status as u16,
+                status_code: output.status,
                 quarry_fingerprint: Some(output.fingerprint.clone()),
             },
         },
